Revert "- refactor package arrangement"
This reverts commit 11acd99dcb346b6218f8a7bab287619e426743cf.
