+++ src/main/java/framework/buttons/ButtonSearch.java
@@ -1,7 +1,7 @@
 package framework.buttons;
 
 import framework.SearchEngineInitializer;
-import interface_adaptors.search_engine_ia.SearchResultsViewModel;
+import interface_adaptors.SearchResultsViewModel;
 import interface_adaptors.playlist_ia.RecordViewModel;
 import interface_adaptors.queue_ia.QueueViewModel;
 import interface_adaptors.search_engine_ia.SearchEngineController;
