+++ src/main/java/framework/items/FriendsCollectionItem.java
@@ -1,7 +1,8 @@
 package framework.items;
 
+import entities.user_entities.User;
 import interface_adaptors.playlist_ia.RecordViewModel;
-import interface_adaptors.search_engine_ia.SearchResultsViewModel;
+import interface_adaptors.SearchResultsViewModel;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
