+++ src/main/java/framework/items/PlaylistCollectionItem.java
@@ -1,8 +1,9 @@
 package framework.items;
 
 
-import interface_adaptors.playlist_ia.PlaylistDTOController;
-import interface_adaptors.search_engine_ia.SearchResultsViewModel;
+import entities.playlist_entities.Playlist;
+import interface_adaptors.PlaylistDTOController;
+import interface_adaptors.SearchResultsViewModel;
 import interface_adaptors.playlist_ia.RecordViewModel;
 
 import javax.imageio.ImageIO;
