+++ src/main/java/interface_adaptors/PlaylistCollectionViewModel.java
@@ -1,13 +1,14 @@
-package interface_adaptors.playlist_ia;
+package interface_adaptors;
 
+import entities.playlist_entities.Playlist;
 import framework.buttons.ButtonCreatePlaylist;
 import framework.items.PlaylistCollectionItem;
-import interface_adaptors.AbstractViewModel;
 import interface_adaptors.user_login_ia.UserStatusObserver;
 import interface_adaptors.user_login_ia.UserStatusViewModel;
 
 import javax.swing.*;
 import java.awt.*;
+import java.util.ArrayList;
 import java.util.List;
 
 public class PlaylistCollectionViewModel extends AbstractViewModel<List<String>> implements UserStatusObserver {
