+++ src/main/java/interface_adaptors/search_engine_ia/SearchEnginePresenter.java
@@ -2,6 +2,10 @@ package interface_adaptors.search_engine_ia;
 
 import abr.search_engine_abr.SearchEngineOutputBoundary;
 import abr.search_engine_abr.SearchEngineResponseModel;
+import interface_adaptors.SearchResultsViewModel;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class SearchEnginePresenter implements SearchEngineOutputBoundary {
